move test_runner.h/cpp to subdirectory
